BLD: Updated `__main__.py`
which makes 'src/dpms' a python executable folder. I am not going to
make it a library (by adding '__init__.py') because it will never be
imported. It will just be executed from the command line.

NOTE: '__main__.py' has required objects (imported, defined, etc.)
(taken from 'src/main.py') that will be required by other objects but
those objects don't depend on the other objects in 'src/main.py'. This
is made in the first commit of the file because it is definitely gonna
use all those in its future development.