DEV: Added `myHelpDeterminer`
Refer comment in code

I have put this piece of code here in order to set the help_order of
every command in the same order they are created in this script. This is
very useful as one does not have to edit every first line of command
declaration when declaring a new command in the begginning or somewhere
in the middle (`@cli.command(help_priority=int)`) just use
`next(myHelpDeterminer)` in the place of `int` to show the command in
the same order in the script. Make sure to edit the value of `total`
(set it to the total number of commands so that we don't get StopIteration
error when iterating more than thelength of range)