BLD: Diff folders for diff versions of the project
Started to maintain different folders for different versions to store
and build separate versions separately without disturbing each other.
Because all the versions are needed (I just now realised it).

(The first version was a script (menu driven program) followed by a CLI
app. The next one will be a website for sure)